BUGFIX: The code for the GrenadeButtonController.cs and the GrenadeThrowHandler.cs scripts were modified so that the regular grenades work perfectly, and so that all the animations now work properly.
